Create a resource map RDF/XML file and save is to a temporary path

This is a convenience wrapper around the constructor of the ResourceMap class from DataPackage.

generate_resource_map(metadata_pid, data_pids = NULL,
  child_pids = NULL, other_statements = NULL,
  resolve_base = "https://cn.dataone.org/cn/v2/resolve",
  resource_map_pid = NULL)

Arguments

metadata_pid
(character) PID of the metadata object.

data_pids
(character) PID(s) of the data objects.

child_pids
(character) Optional. PID(s) of child resource maps.

other_statements
(data.frame) Extra statements to add to the resource map.

resolve_base
(character) Optional. The resolve service base URL.

resource_map_pid
(character) The PID of a resource map.

Value

(character) Absolute path to the resource map on disk.

Examples

# NOT RUN {
generate_resource_map("X", "Y", "Z",
                      other_statements = data.frame(subject="http://example.com/me",
                                                    predicate="http://example.com/foo",
                                                    object="http://example.com/bar"))
# }