Score a metadata document against a MetaDIG suite

This function scores a metadata document against a MetaDIG suite. The default suite is for the Arctic Data Center.

mdq_run(document, suite_id = "arctic.data.center.suite.1")

Arguments

document
(eml/character) Either an EML object or path to a file on disk.

suite_id
(character) Specify a suite ID. Should be one of https://quality.nceas.ucsb.edu/quality/suites.

Value

(data.frame) A sorted data.frame of check results.

Examples

# NOT RUN {
# Check an EML document you are authoring
library(EML)
mdq_run(new("eml"))

# Check an EML document that is saved to disk
mdq_run(system.file("examples", "example-eml-2.1.1.xml", package = "EML"))
# }