editing.R

High-level functions for managing content. Publish an object on a member node

publish_object(mn, path, format_id = NULL, pid = NULL, sid = NULL,
  clone_pid = NULL, public = TRUE)

Arguments

mn
(MNode) The Member Node to publish the object to.

path
the path to the file to be published

format_id
(character) Optional. The format ID to set for the object. When not set, guess_format_id will be used to guess the format ID. Should be a DataONE format ID.

pid
(character) Optional. The PID to use with the object.

sid
(character) Optional. The SID to use with the new object.

clone_pid
(character) PID of objet to clone System Metadata from

Details

Use sensible defaults to publish an object on a member node. If identifier is provided, use it, otherwise generate a UUID. If clone_id is provided, then retrieve the system metadata for that identifier and use it to provide rightsHolder, accessPolicy, and replicationPolicy metadata. Note that this function only uploads the object to the Member Node, and does not add it to a data package, which can be done separately.