Update a data object and associated resource map and metadata

This function updates a data object and then automatically updates the package resource map with the new data PID. If an object already has a dataTable, otherEntity, or spatialVector with a working physical section, the EML will be updated with the new physical. It is a convenience wrapper around update_object() and publish_update().

update_package_object(mn, data_pid, new_data_path, resource_map_pid,
  format_id = NULL, public = TRUE, use_doi = FALSE, ...)

Arguments

mn
(MNode) The Member Node of the data package.

data_pid
(character) PID for data object to update.

new_data_path
(character) Path to new data object.

resource_map_pid
(character) PID for resource map to update.

format_id
(character) Optional. The format ID to set for the object. When not set, guess_format_id() will be used to guess the format ID. Should be a DataONE format ID.

public
(logical) Optional. Make the update public. If FALSE, will set the metadata and resource map to private (but not the data objects). This applies to the new metadata PID and its resource map and data object. Access policies are not affected.

use_doi
(logical) Optional. If TRUE, a new DOI will be minted.

...
Other arguments to pass into publish_update().

Value

(character) Named character vector of PIDs in the data package, including PIDs for the metadata, resource map, and data objects.

See also

Examples

# NOT RUN {
cnTest <- dataone::CNode("STAGING")
mnTest <- dataone::getMNode(cnTest,"urn:node:mnTestARCTIC")

pkg <- create_dummy_package_full(mnTest, title = "My package")

file.create("new_file.csv")
update_package_object(mnTest, pkg$data[1], "new_file.csv", pkg$resource_map, format_id = "text/csv")
file.remove("new_file.csv")
# }