Get an approximate list of the datasets in a user's profile

This function is intended to be (poorly) simulate what a user sees when they browse to their "My Data Sets" page (their #profile URL). It uses a similar Solr to what Metacat UI uses to generate the list. The results of this function may be the same as what's on the #profile page but may be missing some of the user's datasets when:

view_profile(mn, subject, fields = c("identifier", "title"))

Arguments

mn
(MNode) The Member Node to query.

subject
(character) The subject to find the datasets for. This is likely going to be your ORCID, e.g. http://orcid.org....

fields
(character) A vector of Solr fields to return.

Value

(data.frame) data.frame with the results.

Details

- The user has any datasets in their #profile that the person running the query (you) can't read. This is rare on arcticdata.io but possible because arctic-data-admins usually has read/write/changePermission permissions on every object. - The user has datasets owned by an Equivalent Identity of the subject being queried. This is rare, especially on arcticdata.io.

Examples

## Not run: ------------------------------------
# options(...set...your...token....)
# mn <- env_load('production')$mn
# me <- get_token_subject()
# profile(mn, me)
# 
# // Get a custom set of fields
# view_profile(mn, me, "origin")
## ---------------------------------------------